+++ src/QorusDelete.ts
@@ -8,7 +8,7 @@ import { t } from 'ttag';
 
 
 class QorusDelete {
-    private interfaces = {};
+    private interfaces: any = {};
 
     deleteInterfaces(iface_kind: string, ids: string[]) {
         const {ok, active_instance, token} = qorus_request.activeQorusInstanceAndToken();
@@ -24,13 +24,20 @@ class QorusDelete {
                     (({ name, version }) => ({ name, version }))(step)
                 ));
             }
-            return (({ name, version }) => ({ name, version }))(iface);
+            if (iface.name && iface.version) {
+                return (({ name, version }) => ({ name, version }))(iface);
+            } else {
+                return (({ id }) => ({ id }))(iface);
+            }
         });
 
         let iface_post_kind: string;
         switch (iface_kind) {
-            case 'classes': iface_post_kind = 'class'; break;
-            default: iface_post_kind = iface_kind.slice(0, -1);
+            case 'classes':
+                iface_post_kind = 'class';
+                break;
+            default: // only remove 's' to get the singular
+                iface_post_kind = iface_kind.slice(0, -1);
         }
         const config: any = vscode.workspace.getConfiguration('qorusDeployment')
                                 || {reload: false, 'verbosity-level': 1};
@@ -70,7 +77,7 @@ class QorusDelete {
         qorus_request.doRequestAndCheckResult(options, texts, () => {
             qorus_webview.postMessage({
                 action: 'deletion-finished',
-                iface_kind: iface_kind
+                iface_kind
             });
         });
     }
@@ -88,11 +95,16 @@ class QorusDelete {
 
         const isIdKey = (key: string): boolean => {
             switch (iface_kind) {
-                case 'classes':   return key === 'classid';
-                case 'functions': return key === 'function_instanceid';
+                case 'classes':
+                    return key === 'classid';
+                case 'groups':
+                case 'valuemaps':
+                    return key === 'id';
+                case 'pipelines':
+                case 'fsms':
+                    return key === 'name';
                 default:
-                    return key.slice(-2) === 'id'
-                        && key.slice(0, -2) === iface_kind.slice(0, -1);
+                    return key.slice(-2) === 'id' && key.slice(0, -2) === iface_kind.slice(0, -1);
             }
         };