+++ ui-test/src/tests/workflow.ts
@@ -6,6 +6,7 @@ import { openInterfaceFromTreeView } from '../utils/treeView';
 import {
     addNewMultiSelectItemAndSelectIt,
     clickElement,
+    closeLastDialog,
     fillTextField,
     getElementAttribute,
     getElements,
@@ -49,19 +50,15 @@ export const fillsWorkflowFields = async (webview: WebView) => {
     await selectField(webview, 'workflow-autostart');
     await selectField(webview, 'classes');
     await selectNthFilteredDropdownItem(webview, 'name', 'ClassForFSMTest', 1, 2);
-    await selectField(webview, 'constants');
-    await addNewMultiSelectItemAndSelectIt(webview, 'test', 2);
     await selectAndFillField(webview, 'detach', 'test');
     await selectAndFillField(webview, 'errorfunction', 'test');
     await selectAndFillField(webview, 'error_handler', 'test');
-    await selectField(webview, 'functions');
-    await addNewMultiSelectItemAndSelectIt(webview, 'test', 3);
     await selectField(webview, 'groups');
-    await selectMultiselectItemsByNumbers(webview, [1], 4);
+    await selectMultiselectItemsByNumbers(webview, [1], 2);
     await selectField(webview, 'keylist');
-    await addNewMultiSelectItemAndSelectIt(webview, 'test', 5);
+    await addNewMultiSelectItemAndSelectIt(webview, 'test', 3);
     await selectField(webview, 'mappers');
-    await selectMultiselectItemsByNumbers(webview, [1], 6);
+    await selectMultiselectItemsByNumbers(webview, [1], 4);
     await selectAndFillField(webview, 'max_instances', 5);
     await selectAndFillField(webview, 'onetimeinit', 'test');
     await selectField(webview, 'workflow_options');
@@ -76,7 +73,7 @@ export const fillsWorkflowFields = async (webview: WebView) => {
     await fillTextField(webview, 'field-key', 'foo', 2);
     await fillTextField(webview, 'field-value', 'bar', 2);
     await selectField(webview, 'vmaps');
-    await addNewMultiSelectItemAndSelectIt(webview, 'test', 7);
+    await addNewMultiSelectItemAndSelectIt(webview, 'test', 5);
 
     await submitInterface(webview, 'workflow');
 
@@ -115,6 +112,20 @@ export const addsExistingStepFromWorkflowDiagram = async (webview: WebView) => {
     expect(await getElements(webview, 'workflow-diagram-step')).to.have.length(2);
 };
 
+export const canOpenStepForEditingFromDiagram = async (webview: WebView) => {
+    await clickElement(webview, 'workflow-diagram-step', 2);
+    await sleep(500);
+    await clickElement(webview, 'field-step-edit-reference');
+    await sleep(3000);
+
+    expect(await getElementAttribute(webview, 'field-name', 'value')).to.eq('StepForWorkflowTest');
+
+    await closeLastDialog(webview);
+    await sleep(300);
+    await closeLastDialog(webview);
+    await sleep(300);
+};
+
 export const submitsWorkflowAndChecksFiles = async (webview: WebView) => {
     await submitInterface(webview, 'workflow-steps');
 
@@ -132,12 +143,10 @@ export const editsWorkflowAndChecksFiles = async (webview: WebView) => {
     await openInterfaceFromTreeView('WorkflowTest', webview);
     await sleep(4000);
 
-    // Constants, value-maps, functions and keylist should be empty because their interfaces do not exist
+    // value-maps and keylist should be empty because their interfaces do not exist
     // and submit button should be disabled
     expect(await getElementAttribute(webview, 'interface-creator-submit-workflow', 'disabled')).to.eq('true');
 
-    await removeField(webview, 'constants');
-    await removeField(webview, 'functions');
     await removeField(webview, 'vmaps');
 
     await resetAndFillTextField(webview, 'field-name', 'WorkflowTestEdited');