+++ web/containers/InterfaceCreator/index.tsx
@@ -5,6 +5,7 @@ import withInitialDataConsumer from '../../hocomponents/withInitialDataConsumer'
 import withTextContext from '../../hocomponents/withTextContext';
 import ClassConnectionsStateProvider from '../ClassConnectionsStateProvider';
 import { ConnectionView } from './connection';
+import ErrorsView from './errorsView';
 import FSMView from './fsm';
 import LibraryView from './libraryView';
 import MapperView from './mapperView';
@@ -19,13 +20,19 @@ export interface ICreateInterface {
     initialData: any;
     onSubmit: any;
     context: any;
+    data: any;
 }
 
-export const CreateInterface: FunctionComponent<ICreateInterface> = ({ initialData, onSubmit, context }) => {
+export const CreateInterface: FunctionComponent<ICreateInterface> = ({ initialData, onSubmit, data, context }) => {
+    initialData = { ...initialData, ...data };
+
+    const getName: () => string = () =>
+        initialData?.[initialData.subtab]?.name || initialData?.[initialData.subtab]?.path;
+
     return (
         <Box fill style={{ overflow: 'hidden' }}>
             <div className={'fullHeightTabs'}>
-                <Tab type={initialData.subtab}>
+                <Tab name={getName()} type={initialData.subtab} data={initialData}>
                     {initialData.subtab === 'fsm' && (
                         <FSMView fsm={initialData.fsm} onSubmitSuccess={onSubmit} interfaceContext={context} />
                     )}
@@ -143,6 +150,9 @@ export const CreateInterface: FunctionComponent<ICreateInterface> = ({ initialDa
                         </CreatorWrapper>
                     )}
                     {initialData.subtab === 'type' && <TypeView onSubmitSuccess={onSubmit} />}
+                    {initialData.subtab === 'errors' && (
+                        <ErrorsView errors={initialData.errors} onSubmitSuccess={onSubmit} interfaceContext={context} />
+                    )}
                 </Tab>
             </div>
         </Box>