+++ web/containers/InterfaceCreator/panel.tsx
@@ -12,7 +12,7 @@ import {
     reduce,
     size,
     uniqBy,
-    upperFirst,
+    upperFirst
 } from 'lodash';
 import isArray from 'lodash/isArray';
 import React, { FormEvent, FunctionComponent, useContext, useEffect, useRef, useState } from 'react';
@@ -40,7 +40,12 @@ import withFieldsConsumer from '../../hocomponents/withFieldsConsumer';
 import withGlobalOptionsConsumer from '../../hocomponents/withGlobalOptionsConsumer';
 import withInitialDataConsumer from '../../hocomponents/withInitialDataConsumer';
 import withMapperConsumer from '../../hocomponents/withMapperConsumer';
-import withMessageHandler, { TMessageListener, TPostMessage } from '../../hocomponents/withMessageHandler';
+import withMessageHandler, {
+    addMessageListener,
+    postMessage,
+    TMessageListener,
+    TPostMessage
+} from '../../hocomponents/withMessageHandler';
 import withMethodsConsumer from '../../hocomponents/withMethodsConsumer';
 import withStepsConsumer from '../../hocomponents/withStepsConsumer';
 import withTextContext from '../../hocomponents/withTextContext';
@@ -168,8 +173,6 @@ export const ActionsWrapper = styled.div`
 
 const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
     type,
-    addMessageListener,
-    postMessage,
     t,
     fields,
     setFields,
@@ -277,6 +280,13 @@ const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
                 handleFieldChange(field, value);
             }),
         ]);
+
+        return () => {
+            // Remove the current listeners
+            fieldListeners.forEach((listener) => {
+                listener();
+            });
+        };
     }, [fields]);
 
     useEffect(() => {
@@ -531,26 +541,28 @@ const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
     };
 
     const toggleDisableField: (fieldName: string, disabled: boolean) => void = (fieldName, disabled) => {
-        setFields(
-            type,
-            (current: IField[]) =>
-                map(current, (field: IField) => ({
-                    ...field,
-                    disabled: fieldName === field.name ? disabled : field.disabled,
-                })),
-            activeId,
-            interfaceIndex
-        );
-        setSelectedFields(
-            type,
-            (current: IField[]) =>
-                map(current, (field: IField) => ({
-                    ...field,
-                    disabled: fieldName === field.name ? disabled : field.disabled,
-                })),
-            activeId,
-            interfaceIndex
-        );
+        setTimeout(() => {
+            setFields(
+                type,
+                (current: IField[]) =>
+                    map(current, (field: IField) => ({
+                        ...field,
+                        disabled: fieldName === field.name ? disabled : field.disabled,
+                    })),
+                activeId,
+                interfaceIndex
+            );
+            setSelectedFields(
+                type,
+                (current: IField[]) =>
+                    map(current, (field: IField) => ({
+                        ...field,
+                        disabled: fieldName === field.name ? disabled : field.disabled,
+                    })),
+                activeId,
+                interfaceIndex
+            );
+        }, 1000);
     };
 
     const handleAddClick: (fieldName: string) => void = (fieldName) => {
@@ -704,9 +716,28 @@ const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
         if (!onSubmit || forceSubmit) {
             let newData: { [key: string]: any };
             // If this is service methods
-            if (type === 'service-methods' || type === 'mapper-methods') {
-                const intrfType = type === 'service-methods' ? 'service' : 'mapper-code';
-                const subItemType = type === 'service-methods' ? 'methods' : 'mapper-methods';
+            if (type === 'service-methods' || type === 'mapper-methods' || type === 'error') {
+                let intrfType;
+                let subItemType;
+
+                switch (type) {
+                    case 'mapper-methods': {
+                        intrfType = 'mapper-code';
+                        subItemType = 'mapper-methods';
+                        break;
+                    }
+                    case 'error': {
+                        intrfType = 'errors';
+                        subItemType = 'errors_errors';
+                        break;
+                    }
+                    default: {
+                        intrfType = 'service';
+                        subItemType = 'methods';
+                        break;
+                    }
+                }
+
                 // Get the service data
                 newData = reduce(
                     allSelectedFields[intrfType][interfaceIndex],
@@ -751,6 +782,8 @@ const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
                 iface_kind = 'service';
             } else if (type === 'mapper-methods') {
                 iface_kind = 'mapper-code';
+            } else if (type === 'error') {
+                iface_kind = 'errors';
             }
             // Config items use the parent type
             if (parent) {
@@ -779,7 +812,7 @@ const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
                         no_data_return: !!onSubmitSuccess,
                         iface_id: interfaceId,
                     },
-                    t(`Saving ${type}...`)
+                    t(`Saving ${iface_kind}...`)
                 );
             } else {
                 let true_type: string;
@@ -800,16 +833,14 @@ const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
                             default_value_true_type: true_type,
                         },
                         orig_data:
-                            type === 'service-methods'
-                                ? initialData.service
-                                : type === 'mapper-methods'
-                                ? initialData['mapper-code']
+                            type === 'service-methods' || type === 'mapper-methods' || type === 'error'
+                                ? initialData[iface_kind]
                                 : data,
                         open_file_on_success: !onSubmitSuccess && openFileOnSubmit !== false,
                         no_data_return: !!onSubmitSuccess,
                         iface_id: interfaceId,
                     },
-                    t(`Saving ${type}...`)
+                    t(`Saving ${iface_kind}...`)
                 );
             }
             if (result.ok) {
@@ -820,10 +851,6 @@ const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
                 // local fields will be unmounted
                 if (type === 'config-item') {
                     resetFields(type, interfaceIndex);
-                } else {
-                    // Reset the interface data
-                    resetAllInterfaceData(type);
-                    resetClassConnections && resetClassConnections();
                 }
             }
         }
@@ -1062,7 +1089,12 @@ const InterfaceCreatorPanel: FunctionComponent<IInterfaceCreatorPanel> = ({
                     <ActionsWrapper>
                         <ButtonGroup fill>
                             <Tooltip content={t('SelectAllTooltip')}>
-                                <Button text={t('SelectAll')} icon={'plus'} onClick={handleAddAll} />
+                                <Button
+                                    text={t('SelectAll')}
+                                    icon={'plus'}
+                                    onClick={handleAddAll}
+                                    name="add-all-fields"
+                                />
                             </Tooltip>
                         </ButtonGroup>
                     </ActionsWrapper>