+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 2.1.0
+
+#### New features
+
+- new interface: Errors
+- workflows can now have a errors assigned
+- deployment of an interface now (optionally) includes all referenced interfaces
+
+#### Bug fixes
+
+- fixed a bug where the the IDE would crash when opening step for editing from the step diagram
+- it is now properly possible to create a Class from another Class
+- icons representing types of interfaces were unified, now they are the same in the tree view and in the webview
+
 ## 2.0.2
 
 #### New features